Move LSIF jobs pages into enterprise/site-admin. (#6560)
