+++ web/src/enterprise/site-admin/SiteAdmin.scss
@@ -1,0 +1,2 @@
+@import './SiteAdminLsifJobPage';
+@import './SiteAdminLsifJobsPage';
