+++ web/src/enterprise/site-admin/SiteAdminLsifJobPage.scss
