+++ web/src/site-admin/SiteAdmin.scss
@@ -6,8 +6,6 @@
 @import './SiteAdminSurveyResponsesPage';
 @import './SiteAdminAddExternalServicesPage';
 @import './SiteAdminManagementConsolePassword';
-@import './SiteAdminLsifJobPage';
-@import './SiteAdminLsifJobsPage';
 
 .site-admin-area .settings-file {
     flex: 1;
