+++ web/src/site-admin/sidebaritems.ts
@@ -25,10 +25,6 @@ export const primaryGroup: SiteAdminSideBarGroup = {
             label: 'Repositories',
             to: '/site-admin/repositories',
         },
-        {
-            label: 'LSIF Jobs',
-            to: '/site-admin/lsif-jobs',
-        },
     ],
 }
